improved ui generified implementation
