+++ app/src/main/res/values/strings.xml
@@ -792,7 +792,7 @@
     <string name="dialog_field_sort_header_tv">Please add the attributes to sort plots by. Drag items to change the sort order.</string>
     <string name="dialog_field_sort_add_btn_text">Add</string>
     <string name="list_item_field_sort_delete_content_description">Delete the sort attribute.</string>
-    <string name="dialog_field_sort_title">Field Attribute Sort Order</string>
+    <string name="dialog_sort_title">Entry Sort Order</string>
     <string name="list_item_field_sort_drag_content_description">Drag to move item.</string>
     <string name="dialog_field_sort_no_attributes_found">No attributes found.</string>
     <string name="dialog_field_sort_no_more_attributes_found">All attributes have been used.</string>
@@ -800,5 +800,9 @@
     <string name="act_trait_editor_no_traits_exist">No traits exist.</string>
     <string name="preferences_beta_title">Beta</string>
     <string name="dialog_field_sort_revert_btn_text">Default Order</string>
+    <string name="dialog_field_sort_delete_all_content_description">Deletes all sort entries.</string>
+    <string name="dialog_field_sort_toggle_order_content_description">Reverses sort order.</string>
+    <string name="dialog_field_sort_add_content_description">Opens a dialog to add a new attribute to sort by.</string>
+    <string name="dialog_field_sort_title">Field Sort Order</string>
 
 </resources>
